# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Redoubt API
  description: |
    Redoubt is a secure, invite-only communication platform API.

    All error responses conform to RFC 9457 Problem Details for HTTP APIs.
    All timestamps are in ISO 8601 UTC format.
    All IDs are UUID v4.
  version: 0.1.0
  contact:
    name: Redoubt
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-02-22T10:30:00Z"

    # Error schemas (RFC 9457)
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://redoubt.app/problems/validation-error"
        title:
          type: string
          description: Short, human-readable summary
          example: "Validation Failed"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation
          example: "The request body contains invalid fields"
        instance:
          type: string
          description: URI reference for this occurrence
          example: "/api/v1/auth/register"
        request_id:
          type: string
          description: X-Request-ID for tracing
          example: "abc-123-def"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: "username"
        message:
          type: string
          example: "must be between 3 and 32 characters"

    # Auth schemas
...
    Space:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "My Community"
        icon_url:
          type: string
          nullable: true
        owner_id:
          $ref: '#/components/schemas/UUID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    SpaceWithChannels:
      allOf:
        - $ref: '#/components/schemas/Space'
        - type: object
          properties:
            channels:
              type: array
              items:
                $ref: '#/components/schemas/Channel'
    CreateSpaceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9 ]+$"
          example: "My Community"
    UpdateSpaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9 ]+$"
    SpacesListResponse:
      type: object
      properties:
        spaces:
          type: array
          items:
            $ref: '#/components/schemas/Space'

    # Member schemas
    Member:
      type: object
      properties:
        user_id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, admin, member]
        joined_at:
          $ref: '#/components/schemas/Timestamp'
    MembersListResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
    ChangeMemberRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [admin, member]

    # Channel schemas
    Channel:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        space_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "general"
        type:
          type: string
          enum: [text, voice]
        position:
          type: integer
          example: 0
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
paths:
...
  /spaces:
    get:
      tags: [Spaces]
      summary: List spaces
      description: List all spaces the user is a member of.
      operationId: listSpaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of spaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpacesListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Spaces]
      summary: Create space
      description: Create a new space. Requires instance admin role.
      operationId: createSpace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceRequest'
      responses:
        '201':
          description: Space created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceWithChannels'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /spaces/{id}:
    get:
      tags: [Spaces]
      summary: Get space
      description: Get space details. Requires space membership.
      operationId: getSpace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Space details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceWithChannels'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Spaces]
      summary: Update space
      description: Update space details. Requires admin role.
      operationId: updateSpace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpaceRequest'
      responses:
        '200':
          description: Space updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Spaces]
      summary: Delete space
      description: Soft delete a space. Requires owner role.
      operationId: deleteSpace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Space deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /spaces/{id}/members:
    get:
      tags: [Spaces]
      summary: List members
      description: List all members of a space. Requires space membership.
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /spaces/{id}/members/{userId}:
    delete:
      tags: [Spaces]
      summary: Kick member
      description: Remove a member from the space. Requires admin role.
      operationId: kickMember
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Spaces]
      summary: Change member role
      description: Change a member's role. Requires owner role.
      operationId: changeMemberRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeMemberRoleRequest'
      responses:
        '200':
          description: Role changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Channel endpoints
...